move: Load 코드 common 패키지로 이동
